Refactoring package name Make common package that include exception, configuration, model.
