+++ back-end/src/main/java/com/bangstagram/common/exception/CommonControllerAdvice.java
@@ -1,4 +1,4 @@
-package com.bangstagram.exception;
+package com.bangstagram.common.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -12,8 +12,9 @@ import org.springframework.web.bind.annotation.ResponseStatus;
  */
 
 @Slf4j
-@ControllerAdvice(basePackages = {"com.bangstagram.timeline"})
-public class timelineControllerAdvice {
+@ControllerAdvice
+public class CommonControllerAdvice {
+
     @ExceptionHandler(DoNotExistException.class)
     @ResponseStatus(value = HttpStatus.BAD_REQUEST)
     public String requestNotExistId(DoNotExistException e) {
