+++ back-end/src/main/java/com/bangstagram/common/model/CommonEntity.java
@@ -1,4 +1,4 @@
-package com.bangstagram.auditing;
+package com.bangstagram.common.model;
 
 import lombok.Getter;
 import org.springframework.data.annotation.CreatedDate;
@@ -16,7 +16,7 @@ import java.time.LocalDateTime;
 @MappedSuperclass
 @EntityListeners(value = {AuditingEntityListener.class})
 @Getter
-public abstract class TimeAuditorEntity {
+public abstract class CommonEntity {
     @CreatedDate
     private LocalDateTime createdAt;
 }
